sign-req: Option 'comply', certificate subject will comply with CA
If a request file is received with unsuitable subject fields then
force the signed certificate to comply with the subject fields of
the CA certificate.

This is achieved as follows:

In DN mode 'cn_only', only the commonName can be changed. Use global
option --req-cn='new-name' and sign-req command option 'comply' to
force the commanName field to be changed.

In DN mode 'org', all the fields can be changed. Use global options
--req-*='new value' and sign-req command option 'comply' to force
all the fields to comply.  All fields not explicitly specified are
loaded from the vars file, which will be configured in 'org' mode.

The signed certificate output-file will retain the file-name-base
of the original request-file, to maintain compatibility with the
commands 'revoke' and 'renew'.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>