build-ca: Remove specific raw mode code, replaced by standard code
The first version of 'build-ca raw' used completely independent code
to build the CA key and certificate. This was sectioned off by use of
flag EASYRSA_RAW_CA, which allowed easy review of the raw mode code.

The raw mode code is simply a version of the standard code, configured
to build a CA using the SSL library built-in password management code.

Remove the stand-alone raw mode code and replace it by configuring the
standard EasyRSA options to build a CA, using raw mode.

Also, make minor corrections to the verbose messages.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>