export-p12: Always set 'friendlyName' to file-name-base
This changes default behavior from NEVER setting a friendlyName
to ALWAYS setting friendlyName to the file-name-base.

The friendlyName is "typically displayed in list boxes by software
importing the file".

Having Easy-RSA populate this field with known values, which do not
leak excess information, is considered to be appropriate.

An alternate friendlyName can be configured by using:
* Global option '--usefn=<friendlyName>'

Fallback to previous behavior can be configured by using:
* Command option 'nofn' ('friendlyname' will not be set)

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>