Always try to use built-in openssl-easysa.cnf here-doc
The SSL config file is always verified as present when
secure_session() is called, at the beginning of execution.

This change forces the use of a temp-file, containing the
current here-doc version of the SSL config file.

If the current, existing file is NOT recognised as a hash
identical file to either the release version or the here-doc
version, which can include new version hashes, then the file
is left in place and used.

Otherwise, if the current file is recognised then it will be
deleted and, from then on, a temp-file based on the here-doc
will always be used in future.

This approach also solves the problem of the unpopulated SSL
config file field 'default_days = $ENV::EASYRSA_CERT_EXPIRE'.

The solution here is to remove the unnecessary 'default_days'
definition. The 'default_days' field is always configured via
the command line.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>