write: Always use verify_working_env()
Move the creation of secure_session and openssl-easyrsa.cnf (Temp)
to verify_working_env(). Create a session and temp-file in the temp-dir
EASRSA_TEMP_DIR, provided that the directory exists.

By default, EASYRSA_TEMP_DIR is set to the current PKI but that can be
decoupled and command 'write' can be run without a PKI, so long as a
valid temp-dir exists.

Correction to error message about 'write <type>' being unknown.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>