Make keyUsage critical in code-signing
This is required, as described in:

"Baseline Requirements for the Issuance and Management of
Publicly-Trusted TLS Server certificates v2.0.2", CA/Browser Forum.

See in particular table 7.1.3.2.1 where keyUsage is marked CRITICAL.

This change is required for easy-rsa certificates to be passed as
valid code-signing certificates by OpenSSL 3.2.0, which gives the
above as its source.

Signed-off-by: Ed Smith <ed.smith@collabora.com>