Remove escape_hazard()
This was an ill conceived atempt, to auto-escape characters used
in the Organizational fields of an extended distinguished name.

The original problem was due to using 'sed', to expand the SSL
config file for use by LibreSSL, only.

Use of 'sed' has been replaced by 'here-doc' expansion, which
means 'escape_hazard()' is pointless.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>