source_vars(): Correct regex guard - Remove unnecessary guard
The regex used by grep incorrectly expects a [:blank:] before 'export'.
This allows a line beginning with 'export' to slip through the net.
The same goes for 'unset'

Also, remove regex for '  > ', accidental --verbose output, copied to
the vars file.  --verbose no lomger uses '>' as a line label, instead
--verbose output is labelled with '#', which would be interpretted as
a comment in an input file.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>