easyrsa_mkdir(): Remove use of 'mkdir -p', use only 'mkdir'
Windows 11 fails to execute 'mkdir.exe -p $foo' and fails
to return an error.

easyrsa_mkdir() is a simple wrapper function for 'mkdir',
which specifically checks that the requested directory is
created, without relying on the exit status of 'mkdir.exe'.

easyrsa_mkdir() does not support the '-p' (Parent) switch.
Instead, `easyrsa` is tasked with creating the parent dirs
as required.

The old easyrsa_mkdir_p() is removed and replaced.

This is not a fix for Windows 11, it is addressing a known
issue by ensuring the failure is captured correctly.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>