Remove --old-safe-ssl $EASYRSA_LEGACY_SAFE_SSL
This switch was used to allow 'sed' expansion of openssl-easyrsa.cnf
to a "safe" config file, instead of the preferred here-doc expansion.

This is no longer required.

Instead of expanding the SSL config file at the last moment, before
calling the SSL command to complete the EasyRSA command, expansion
takes place at the beginning, when the file is first found/created.

In the case of 'found', it is required to use 'sed' to expand any
file which is not recognised by SHA256 hash. If recognised then the
file in place is ignored and a here-doc expansion is used to create
a Safe SSL config.

In the case of created, that is always done from here-doc expansion.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>