revoke: Make check for conflicting files less intrusive
Ref: commit 38bf2d8c52f706fdc87caf888c60c2db0e28ca2d

In order to avoid accidentally revoking an issued certificate,
when an expired or renewed certificate is intended, guard command
'revoke' by checking for certificates of the same file-name in
the 'expired' or 'renewed/issued' directories, first.

Move the check to the revoke() function.

Also, pass the target directory as the first argument to the
revoke() function, instead of setting a variable in the command
selection phase.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>