Introduce "robust" lock-file mechanism
The intention of "robust" is to extend the atomic-nature of
the call to create the lock-file to include failures.

When create_lock_file() fails then $create_lock_file_error is
set. Subsequent checks for lock-file existence are over-ridden.
Thus, if an existing lock-file is removed, during the current
processes lifetime, then easyrsa still maintains the status of
the original lock-file that was tested for.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>